pyproject.toml
[tool.poetry]
name = "mlops-project"
version = "0.1.0"
description = "Project for Applied MLops course"
authors = ["Utnasun Muniev <uamuniev@edu.hse.ru>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.1.4"
numpy = "^1.26.2"
matplotlib = "^3.8.2"
pre-commit = "^3.6.0"
dvc = "^3.36.1"
dvc-gdrive = "^3.0.0"
scikit-learn = "^1.3.2"
catboost = "^1.2.2"
hydra-core = "^1.3.2"
mlflow = "^2.9.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
